Add Cloud Agent environment setup for Firefly III
Configure PHP 8.4, MariaDB, Composer, and Node via Dockerfile and
install/start scripts. Fix portfolio migration foreign key type to match
user_groups.id (bigint).

+++ database/migrations/2026_04_21_000001_create_portfolio_tables.php
@@ -12,7 +12,7 @@ return new class () extends Migration {
         Schema::create('portfolio_accounts', static function (Blueprint $table): void {
             $table->bigIncrements('id');
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('user_group_id');
+            $table->unsignedBigInteger('user_group_id');
             $table->string('name', 255);
             $table->string('platform', 50);
             $table->boolean('active')->default(true);
